Add STX support for Ledger accounts
