+++ src/app/screens/ledger/importLedgerAccount/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import styled, { useTheme } from 'styled-components';
-import { animated, useTransition } from '@react-spring/web';
+import { useTheme } from 'styled-components';
+import { useTransition } from '@react-spring/web';
 import Transport from '@ledgerhq/hw-transport-webusb';
 import ActionButton from '@components/button';
 import {
@@ -17,7 +17,6 @@ import LedgerAddressComponent from '@components/ledger/ledgerAddressComponent';
 import useWalletSelector from '@hooks/useWalletSelector';
 import LedgerInput from '@components/ledger/ledgerInput';
 import FullScreenHeader from '@components/ledger/fullScreenHeader';
-import Switch from 'react-switch';
 import warningIcon from '@assets/img/Warning_red.svg';
 
 import LedgerImportStartSVG from '@assets/img/ledger/ledger_import_start.svg';
@@ -31,312 +30,53 @@ import CheckCircleSVG from '@assets/img/ledger/check_circle.svg';
 import LedgerAccountSwitchSVG from '@assets/img/ledger/account_switch.svg';
 import ArrowLeftIconSVG from '@assets/img/ledger/arrow_left_icon.svg';
 import LedgerFailView from '@components/ledger/failLedgerView';
+import LedgerAssetSelectCard from '@components/ledger/ledgerAssetSelectCard';
 import LedgerConnectionView from '../../../components/ledger/connectLedgerView';
+import {
+  AddAccountNameContainer,
+  AddAccountNameTitleContainer,
+  AddAddressDetailsContainer,
+  AddAddressHeaderContainer,
+  AddressAddedContainer,
+  AssetSelectionButton,
+  AssetSelectionButtonText,
+  ButtonContainer,
+  ConfirmationStep,
+  ConfirmationStepsContainer,
+  ConfirmationText,
+  Container,
+  CreateAnotherAccountContainer,
+  CreateMultipleAccountsText,
+  CustomLink,
+  CustomSwitch,
+  EndScreenContainer,
+  EndScreenTextContainer,
+  ImportBeforeStartContainer,
+  ImportBeforeStartText,
+  ImportBeforeStartTitle,
+  ImportCardContainer,
+  ImportStartContainer,
+  ImportStartImage,
+  ImportStartText,
+  ImportStartTitle,
+  OnBoardingActionsContainer,
+  OnBoardingContentContainer,
+  Option,
+  OptionIcon,
+  OptionsContainer,
+  SelectAssetFootNote,
+  SelectAssetText,
+  SelectAssetTextContainer,
+  SelectAssetTitle,
+  TogglerContainer,
+  TogglerText,
+  WarningIcon,
+} from './index.styled';
 
 const LINK_TO_LEDGER_ACCOUNT_ISSUE_GUIDE =
   'https://support.xverse.app/hc/en-us/articles/17901278165773';
 const LINK_TO_LEDGER_PASSPHRASE_GUIDE =
   'https://support.xverse.app/hc/en-us/articles/17901278165773';
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  flex: 1;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-const OnBoardingContentContainer = styled(animated.div)((props) => ({
-  display: 'flex',
-  width: '100%',
-  flexDirection: 'column',
-  flex: 1,
-  justifyContent: props.className === 'center' ? 'center' : 'none',
-  paddingLeft: props.theme.spacing(8),
-  paddingRight: props.theme.spacing(8),
-}));
-const OnBoardingActionsContainer = styled.div((props) => ({
-  width: '100%',
-  display: 'flex',
-  alignItems: 'flex-start',
-  justifyContent: 'space-between',
-  paddingLeft: props.theme.spacing(8),
-  paddingRight: props.theme.spacing(8),
-  marginBottom: props.theme.spacing(30),
-}));
-const ImportStartImage = styled.img((props) => ({
-  marginLeft: props.theme.spacing(0),
-}));
-
-const ImportStartContainer = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  maxWidth: '328px',
-});
-
-const ImportBeforeStartContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  maxWidth: '328px',
-  paddingTop: props.theme.spacing(16),
-}));
-
-const ButtonContainer = styled.div((props) => ({
-  marginLeft: 3,
-  marginRight: 3,
-  marginTop: props.theme.spacing(4),
-  width: '100%',
-}));
-
-const ImportStartTitle = styled.h1((props) => ({
-  ...props.theme.headline_s,
-  marginTop: props.theme.spacing(40),
-  textAlign: 'center',
-}));
-const ImportStartText = styled.p((props) => ({
-  ...props.theme.body_m,
-  textAlign: 'center',
-  marginTop: props.theme.spacing(6),
-  color: props.theme.colors.white[200],
-}));
-
-const ImportBeforeStartTitle = styled.h1((props) => ({
-  ...props.theme.headline_s,
-  marginTop: props.theme.spacing(5),
-  textAlign: 'left',
-  alignSelf: 'flex-start',
-}));
-const ImportBeforeStartText = styled.p((props) => ({
-  ...props.theme.body_m,
-  textAlign: 'left',
-  marginTop: props.theme.spacing(6),
-  color: props.theme.colors.white[200],
-}));
-
-const ImportCardContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: props.theme.spacing(6),
-}));
-
-const SelectAssetTextContainer = styled.div((props) => ({
-  marginTop: props.theme.spacing(20),
-  marginBottom: props.theme.spacing(16),
-  display: 'flex',
-  flexDirection: 'column',
-  gap: props.theme.spacing(6),
-}));
-
-const SelectAssetTitle = styled.h1((props) => ({
-  ...props.theme.headline_s,
-}));
-
-const SelectAssetText = styled.p((props) => ({
-  ...props.theme.body_m,
-  color: props.theme.colors.white[200],
-  textAlign: 'center',
-}));
-
-const SelectAssetFootNote = styled.p((props) => ({
-  ...props.theme.body_xs,
-  color: props.theme.colors.white[200],
-  marginBottom: props.theme.spacing(25),
-  marginTop: props.theme.spacing(12),
-}));
-
-const AddAddressHeaderContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  gap: props.theme.spacing(8),
-  marginTop: props.theme.spacing(20),
-  marginBottom: props.theme.spacing(8),
-}));
-
-const CreateAnotherAccountContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  gap: props.theme.spacing(8),
-  paddingTop: props.theme.spacing(90),
-  marginBottom: props.theme.spacing(16),
-}));
-
-const AddAddressDetailsContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'flex-start',
-  gap: props.theme.spacing(20),
-}));
-
-const AddressAddedContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  gap: 8px;
-  > :first-child {
-    margin-bottom: 26px;
-  }
-`;
-
-const AddAccountNameContainer = styled.div((props) => ({
-  marginTop: props.theme.spacing(20),
-  width: '100%',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: props.theme.spacing(6),
-}));
-
-const AddAccountNameTitleContainer = styled.div((props) => ({
-  width: '100%',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: props.theme.spacing(4),
-  marginBottom: props.theme.spacing(22),
-}));
-
-const CreateMultipleAccountsText = styled.h3((props) => ({
-  ...props.theme.body_l,
-  textAlign: 'center',
-}));
-
-const EndScreenContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-}));
-
-const EndScreenTextContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  textAlign: 'center',
-  gap: props.theme.spacing(6),
-  marginBottom: props.theme.spacing(20),
-}));
-
-const AssetSelectionButton = styled.button((props) => ({
-  position: 'absolute',
-  left: props.theme.spacing(105),
-  top: props.theme.spacing(60),
-  backgroundColor: 'transparent',
-  display: 'flex',
-  flexDirection: 'row',
-  alignItems: 'center',
-  gap: props.theme.spacing(3),
-}));
-
-const AssetSelectionButtonText = styled.p((props) => ({
-  ...props.theme.body_m,
-  color: props.theme.colors.white[0],
-}));
-
-const ConfirmationText = styled.p((props) => ({
-  fontSize: '0.875rem',
-  fontWeight: 500,
-  textAlign: 'center',
-  marginTop: props.theme.spacing(50),
-}));
-
-const ConfirmationStepsContainer = styled.div((props) => ({
-  display: 'flex',
-  justifyContent: 'center',
-  marginTop: props.theme.spacing(12),
-}));
-
-const OptionsContainer = styled.div((props) => ({
-  width: '100%',
-  marginTop: props.theme.spacing(16),
-}));
-
-interface OptionProps {
-  selected?: boolean;
-}
-const Option = styled.div<OptionProps>((props) => ({
-  width: '100%',
-  backgroundColor: '#21253C',
-  padding: props.theme.spacing(8),
-  paddingTop: props.theme.spacing(7),
-  paddingBottom: props.theme.spacing(7),
-  borderRadius: props.theme.radius(2),
-  fontSize: '0.75rem',
-  marginBottom: props.theme.spacing(6),
-  border: `1px solid ${props.selected ? 'rgba(115, 131, 255, 0.40)' : 'transparent'}`,
-  cursor: 'pointer',
-  userSelect: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  transition: 'border 0.2s ease',
-}));
-
-const OptionIcon = styled.div<OptionProps>((props) => ({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  width: 15,
-  height: 15,
-  borderRadius: '50%',
-  border: `1px solid ${props.theme.colors.white[0]}`,
-  marginRight: props.theme.spacing(10),
-  flex: 'none',
-  '&::after': {
-    content: '""',
-    display: props.selected ? 'block' : 'none',
-    width: 8,
-    height: 8,
-    borderRadius: 100,
-    backgroundColor: props.theme.colors.white[0],
-  },
-}));
-
-interface ConfirmationStepProps {
-  isCompleted: boolean;
-}
-const ConfirmationStep = styled.div<ConfirmationStepProps>((props) => ({
-  width: 32,
-  height: 4,
-  backgroundColor: props.isCompleted ? props.theme.colors.white[0] : props.theme.colors.white[900],
-  borderRadius: props.theme.radius(1),
-  transition: 'background-color 0.2s ease',
-  ':first-child': {
-    marginRight: props.theme.spacing(4),
-  },
-}));
-
-const CustomSwitch = styled(Switch)`
-  .react-switch-handle {
-    background-color: ${({ checked }) =>
-      checked ? '#FFFFFF' : 'rgba(255, 255, 255, 0.2)'} !important;
-    border: ${({ checked }) => (checked ? '' : '4px solid rgba(255, 255, 255, 0.2)')} !important;
-  }
-`;
-
-const TogglerContainer = styled.div((props) => ({
-  display: 'flex',
-  alignItems: 'center',
-  marginTop: props.theme.spacing(18),
-}));
-
-const TogglerText = styled.p((props) => ({
-  marginLeft: props.theme.spacing(8),
-  fontWeight: 500,
-  fontSize: '0.875rem',
-  lineHeight: '140%',
-}));
-
-const CustomLink = styled.a((props) => ({
-  color: props.theme.colors.orange_main,
-}));
-
-const WarningIcon = styled.img({
-  width: 32,
-  height: 32,
-  alignSelf: 'flex-start',
-});
 
 interface Credential {
   publicKey: string;
@@ -490,18 +230,17 @@ function ImportLedger(): JSX.Element {
       }
     }
 
-    if (currentStepIndex === 0) {
-      setCurrentStepIndex(0.5);
+    if (currentStepIndex === 1) {
+      setCurrentStepIndex(1.5);
       return;
     }
 
-    if (currentStepIndex === 0.5) {
-      setCurrentStepIndex(0.75);
+    if (currentStepIndex === 1.5) {
+      setCurrentStepIndex(1.75);
       return;
     }
 
-    // Skip choosing BTC / STX account
-    if (currentStepIndex === 0.75) {
+    if (currentStepIndex === 1.75) {
       setCurrentStepIndex(2);
       return;
     }
@@ -746,7 +485,36 @@ function ImportLedger(): JSX.Element {
                 <ImportStartText>{t('LEDGER_IMPORT_1_SUBTITLE')}</ImportStartText>
               </ImportStartContainer>
             )}
-            {currentStepIndex === 0.5 && (
+            {currentStepIndex === 1 && (
+              <div>
+                <SelectAssetTextContainer>
+                  <SelectAssetTitle>{t('LEDGER_IMPORT_2_TITLE')}</SelectAssetTitle>
+                  <SelectAssetText>{t('LEDGER_IMPORT_2_SUBTITLE')}</SelectAssetText>
+                </SelectAssetTextContainer>
+
+                <ImportCardContainer id="card_container">
+                  <LedgerAssetSelectCard
+                    icon={BtcOrdinalsIconSVG}
+                    title={t('LEDGER_IMPORT_2_SELECT.BTC_TITLE')}
+                    text={t('LEDGER_IMPORT_2_SELECT.BTC_SUBTITLE')}
+                    id="btc_select_card"
+                    isChecked={isBitcoinSelected}
+                    onChange={handleAssetSelect}
+                  />
+
+                  <LedgerAssetSelectCard
+                    icon={StxIconSVG}
+                    title={t('LEDGER_IMPORT_2_SELECT.STACKS_TITLE')}
+                    text={t('LEDGER_IMPORT_2_SELECT.STACKS_SUBTITLE')}
+                    id="stx_select_card"
+                    isChecked={isStacksSelected}
+                    onChange={handleAssetSelect}
+                  />
+                  <SelectAssetFootNote>{t('LEDGER_IMPORT_2_FOOTNOTE')}</SelectAssetFootNote>
+                </ImportCardContainer>
+              </div>
+            )}
+            {currentStepIndex === 1.5 && (
               <ImportBeforeStartContainer>
                 <ImportBeforeStartTitle>
                   {t('LEDGER_BEFORE_GETTING_STARTED.TITLE')}
@@ -772,7 +540,7 @@ function ImportLedger(): JSX.Element {
                 </OptionsContainer>
               </ImportBeforeStartContainer>
             )}
-            {currentStepIndex === 0.75 && (
+            {currentStepIndex === 1.75 && (
               <ImportBeforeStartContainer>
                 <WarningIcon src={warningIcon} alt="Warning" />
                 <ImportBeforeStartTitle>
@@ -850,37 +618,6 @@ function ImportLedger(): JSX.Element {
                 </TogglerContainer>
               </ImportBeforeStartContainer>
             )}
-            {/* Skip choosing BTC / STX account */}
-
-            {/* {currentStepIndex === 1 && (
-              <div>
-                <SelectAssetTextContainer>
-                  <SelectAssetTitle>{t('LEDGER_IMPORT_2_TITLE')}</SelectAssetTitle>
-                  <SelectAssetText>{t('LEDGER_IMPORT_2_SUBTITLE')}</SelectAssetText>
-                </SelectAssetTextContainer>
-
-                <ImportCardContainer id="card_container">
-                  <LedgerAssetSelectCard
-                    icon={BtcOrdinalsIconSVG}
-                    title={t('LEDGER_IMPORT_2_SELECT.BTC_TITLE')}
-                    text={t('LEDGER_IMPORT_2_SELECT.BTC_SUBTITLE')}
-                    id="btc_select_card"
-                    isChecked={isBitcoinSelected}
-                    onChange={handleAssetSelect}
-                  />
-
-                  <LedgerAssetSelectCard
-                    icon={StxIconSVG}
-                    title={t('LEDGER_IMPORT_2_SELECT.STACKS_TITLE')}
-                    text={t('LEDGER_IMPORT_2_SELECT.STACKS_SUBTITLE')}
-                    id="stx_select_card"
-                    isChecked={isStacksSelected}
-                    onChange={handleAssetSelect}
-                  />
-                  <SelectAssetFootNote>{t('LEDGER_IMPORT_2_FOOTNOTE')}</SelectAssetFootNote>
-                </ImportCardContainer>
-              </div>
-            )} */}
             {currentStepIndex === 2 && (
               <LedgerConnectionView
                 title={t(
@@ -1047,27 +784,27 @@ function ImportLedger(): JSX.Element {
             {currentStepIndex === 0 && (
               <ActionButton onPress={handleClickNext} text={t('LEDGER_IMPORT_1_BUTTON')} />
             )}
-            {currentStepIndex === 0.5 && (
+            {currentStepIndex === 1 && (
+              <ActionButton
+                onPress={handleClickNext}
+                text={t('LEDGER_IMPORT_CONTINUE_BUTTON')}
+                disabled={!isBitcoinSelected && !isStacksSelected}
+              />
+            )}
+            {currentStepIndex === 1.5 && (
               <ActionButton
                 onPress={handleClickNext}
                 text={t('LEDGER_IMPORT_CONTINUE_BUTTON')}
                 disabled={selectedLedgerLiveOption === null}
               />
             )}
-            {currentStepIndex === 0.75 && (
+            {currentStepIndex === 1.75 && (
               <ActionButton
                 onPress={handleClickNext}
                 text={t('LEDGER_IMPORT_CONTINUE_BUTTON')}
                 disabled={!isTogglerChecked}
               />
             )}
-            {/* {currentStepIndex === 1 && (
-              <ActionButton
-                onPress={handleClickNext}
-                text={t('LEDGER_IMPORT_CONTINUE_BUTTON')}
-                disabled={!isBitcoinSelected && !isStacksSelected}
-              />
-            )} */}
             {currentStepIndex === 2 && (
               <ActionButton
                 processing={isButtonDisabled}
